// src/types/instance.ts
export enum VersionTypes {
    release = "release",
    snapshot = "snapshot",
    oldBeta = "old_beta",
    oldAlpha = "old_alpha",
}

export enum ModLoaderTypes {
    Vanilla = "Vanilla",
    Forge = "Forge",
    Fabric = "Fabric",
    NeoForge = "Neoforge",
    LiteLoader = "Liteloader",
    Qulit = "Qulit",
}

export enum AddonTypes {
    Optifine = 'Optifine',
    Fabricapi = 'Fabricapi',
    Optifabric = 'Optifabric',
    liteloaderapi = 'Liteloaderapi'
}